refactor(dao): service에서 사용되지 않는 dao Spring Bean 해제 및 테스트 코드 수정
