+++ README.md
@@ -7,7 +7,7 @@
 ---
 
 ![Generic badge](https://img.shields.io/badge/Level2-subway_map-green.svg)
-![Generic badge](https://img.shields.io/badge/test-32_passed-blue.svg)
+![Generic badge](https://img.shields.io/badge/test-41_passed-blue.svg)
 ![Generic badge](https://img.shields.io/badge/version-1.0.0-brightgreen.svg)
 
 > 우아한테크코스 웹 백엔드 4기, 지하철 노선 저장소입니다.
