+++ src/main/java/wooteco/subway/ui/StationController.java
@@ -13,9 +13,10 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import wooteco.subway.domain.Station;
-import wooteco.subway.dto.StationRequest;
-import wooteco.subway.dto.StationResponse;
 import wooteco.subway.service.StationService;
+import wooteco.subway.service.dto.StationServiceRequest;
+import wooteco.subway.ui.dto.StationRequest;
+import wooteco.subway.ui.dto.StationResponse;
 
 @RestController
 @RequestMapping("/stations")
@@ -29,8 +30,10 @@ public class StationController {
 
     @PostMapping
     public ResponseEntity<StationResponse> createStation(@RequestBody StationRequest stationRequest) {
-        final Station newStation = stationService.save(stationRequest);
+        final Station newStation = stationService.save(new StationServiceRequest(stationRequest.getName()));
+
         StationResponse stationResponse = new StationResponse(newStation.getId(), newStation.getName());
+
         return ResponseEntity.created(URI.create("/stations/" + newStation.getId())).body(stationResponse);
     }
