+++ src/main/java/wooteco/subway/ui/dto/LineResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.ui.dto;
 
 import java.util.ArrayList;
 import java.util.List;
