+++ src/main/java/wooteco/subway/ui/dto/StationResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dto;
+package wooteco.subway.ui.dto;
 
 public class StationResponse {
     private Long id;
