--- src/test/java/wooteco/subway/SubwayApplicationTests.java
@@ -1,13 +1,0 @@
-package wooteco.subway;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SubwayApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
